performance_tests/requirements.txt
locustio==0.9.0
beautifulsoup4==4.6.3
